CRUD tests / 4 parts are up, some fail

+++ src/main/java/com/norab/photo/PhotoDao.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 public interface PhotoDao<Photo> {
     List<Photo> selectPhotos();
 
-    int insertPhoto(Photo photo);
+    long insertPhoto(Photo photo);
 
     int deletePhoto(Long id);
